chain: remove block::block
